Reorganize metrics by data modality
