+++ sdmetrics/detection/tabular.py
@@ -1,0 +1,6 @@
+from sdmetrics.multivariate.detection.tabular import LogisticDetector, SVCDetector
+
+__all__ = [
+    'LogisticDetector',
+    'SVCDetector',
+]
