+++ sdmetrics/multivariate/__init__.py
@@ -1,0 +1,8 @@
+from sdmetrics.multivariate import constraint, detection, efficacy, statistical
+
+__all__ = [
+    'constraint',
+    'detection',
+    'efficacy',
+    'statistical'
+]
