+++ sdmetrics/multivariate/detection/tabular/base.py
