+++ sdmetrics/multivariate/statistical/bivariate/discrete.py
@@ -2,9 +2,9 @@
 import numpy as np
 from scipy.special import rel_entr
 
+from sdmetrics.multivariate.statistical.bivariate.base import BivariateMetric
+from sdmetrics.multivariate.statistical.utils import frequencies
 from sdmetrics.report import Goal
-from sdmetrics.statistical.bivariate.base import BivariateMetric
-from sdmetrics.statistical.utils import frequencies
 
 
 class DiscreteDivergence(BivariateMetric):
