+++ sdmetrics/multivariate/statistical/univariate/kstest.py
@@ -1,8 +1,8 @@
 import numpy as np
 from scipy.stats import ks_2samp
 
+from sdmetrics.multivariate.statistical.univariate.base import UnivariateMetric
 from sdmetrics.report import Goal
-from sdmetrics.statistical.univariate.base import UnivariateMetric
 
 
 class KSTest(UnivariateMetric):
